fix(delivery): address quote review feedback
