+++ apps/delivery/package.json
@@ -22,6 +22,7 @@
     },
     "dependencies": {
         "@googlemaps/js-api-loader": "2.1.1",
+        "@vercel/firewall": "1.2.1",
         "next": "16.3.0-preview.6",
         "react": "19.2.7",
         "react-dom": "19.2.7",
